Main Project/backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
marshmallow==3.19.0
psycopg2-binary==2.9.7
gunicorn==20.1.0
python-dotenv==1.0.0
scikit-learn==1.2.2
pandas==2.1.2
joblib==1.2.0
alembic==1.11.1
